Backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
boto3>=1.28.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
gitpython>=3.1.30
httpx>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
numpy>=1.24.0
tqdm>=4.64.0
requests>=2.28.0
PyYAML>=6.0
tenacity>=8.0.0
cachetools>=5.0.0
rich>=13.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
